[jena] Adding JSON-LD Writer instance
Tested with:
    val r = jsonldCompactedWriter.write( graph, os, endpoint )

Comments and code syntax enhancements:

Added some documentation for Generic Parameters, not with @tparam but by a reference in the text.
	NOTE: other @tparam remain elsewhere.
Fix typo in class definition of Transactor regarding typeclass name A .
in JenaModule and JenaRDFWriter, remove useless package path org.w3.banana.io ; add a space after the commas.

Updaded the pre-commit message.